Chapter 8/requirements.txt
langchain==0.0.153
openai==0.27.6
python-dotenv==1.0.0
google-search-results==2.4.2